+++ src/Control/Applicative/Trans/Free.hs
@@ -52,7 +52,9 @@ import Control.Monad (liftM)
 import Data.Functor.Apply
 import Data.Functor.Identity
 import Data.Typeable
-import Data.Monoid
+#if !(MIN_VERSION_base(4,8,0))
+import Data.Monoid (Monoid)
+#endif
 import qualified Data.Foldable as F
 
 -- | The free 'Applicative' for a 'Functor' @f@.