+++ src/Control/Monad/Trans/Iter.hs
@@ -60,6 +60,7 @@ module Control.Monad.Trans.Iter
   , liftIter
   , cutoff
   , never
+  , untilJust
   , interleave, interleave_
   -- * Consuming iterative monads
   , retract
@@ -72,6 +73,7 @@ module Control.Monad.Trans.Iter
   ) where
 
 import Control.Applicative
+import Control.Monad.Catch (MonadCatch(..), MonadThrow(..))
 import Control.Monad (ap, liftM, MonadPlus(..), join)
 import Control.Monad.Fix
 import Control.Monad.Trans.Class
@@ -268,6 +270,14 @@ instance Monad m => MonadFree Identity (IterT m) where
   wrap = IterT . return . Right . runIdentity
   {-# INLINE wrap #-}
 
+instance MonadThrow m => MonadThrow (IterT m) where
+  throwM = lift . throwM
+  {-# INLINE throwM #-}
+
+instance MonadCatch m => MonadCatch (IterT m) where
+  catch (IterT m) f = IterT $ liftM (fmap (`catch` f)) m `catch` (runIterT . f)
+  {-# INLINE catch #-}
+
 -- | Adds an extra layer to a free monad value.
 --
 -- In particular, for the iterative monad 'Iter', this makes the
@@ -308,6 +318,20 @@ liftIter = hoistIterT (return . runIdentity)
 -- | A computation that never terminates
 never :: (Monad f, MonadFree f m) => m a
 never = delay never
+
+-- | Repeatedly run a computation until it produces a 'Just' value.
+-- This can be useful when paired with a monad that has side effects.
+--
+-- For example, we may have @genId :: IO (Maybe Id)@ that uses a random
+-- number generator to allocate ids, but fails if it finds a collision.
+-- We can repeatedly run this with
+--
+-- @
+-- 'retract' ('untilJust' genId) :: IO Id
+-- @
+untilJust :: (Monad m) => m (Maybe a) -> IterT m a
+untilJust f = maybe (delay (untilJust f)) return =<< lift f
+{-# INLINE untilJust #-}
 
 -- | Cuts off an iterative computation after a given number of
 -- steps. If the number of steps is 0 or less, no computation nor